Load token and attempt login on app load

+++ server_instance/shared/utilities/securityToken.js
@@ -1,18 +1,21 @@
 // Store security token in browser
 export function saveToken(securityToken, keepSignedIn) {
-	localStorage.setItem("securityToken", JSON.stringify(securityToken));
-	return JSON.parse(localStorage.getItem("securityToken")) === securityToken;
+	if (keepSignedIn === true) {
+		localStorage.setItem("securityToken", JSON.stringify(securityToken));
+	} else {
+		sessionStorage.setItem("securityToken", JSON.stringify(securityToken));
+	}
 }
 
 // Clear security token from browser
 export function clearToken() {
 	localStorage.removeItem("securityToken");
+	sessionStorage.removeItem("securityToken");
 	localStorage.clear();
 	sessionStorage.clear();
-	return JSON.parse(localStorage.getItem("securityToken")) === null;
 }
 
 // Retrieve security token from browser
 export function getToken() {
-	return JSON.parse(localStorage.getItem("securityToken"));
+	return JSON.parse(sessionStorage.getItem("securityToken") || localStorage.getItem("securityToken"));
 }
